# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SpendingInput {
  /** Amount in Korean Won (KRW) */
  amount: number;
  /** Spending category e.g. 음식, 쇼핑, 오락, 뷰티, 교통, 카페 */
  category: string;
  /** Emotion at time of purchase e.g. 스트레스, 슬픔, 기쁨, 외로움, 무료함, 보상 */
  emotion: string;
  /** Free-text reason for spending */
  reason: string;
  /**
     * Optional additional note
     * @nullable
     */
  note?: string | null;
}

export interface EmotionalAnalysis {
  /** Identified spending pattern in Korean */
  spendingPattern: string;
  /** Emotional trigger analysis in Korean */
  emotionalTrigger: string;
  /** Personalized advice in Korean */
  personalizedAdvice: string;
}

export interface SpendingRecord {
  id: number;
  amount: number;
  category: string;
  emotion: string;
  reason: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  analysis?: EmotionalAnalysis;
}

export interface BreakdownItem {
  label: string;
  count: number;
  amount: number;
}

export interface WeeklyTrendItem {
  week: string;
  amount: number;
  count: number;
}

export interface SpendingSummary {
  totalAmount: number;
  totalCount: number;
  topEmotion: string;
  topCategory: string;
  emotionBreakdown: BreakdownItem[];
  categoryBreakdown: BreakdownItem[];
  weeklyTrend: WeeklyTrendItem[];
  /** @nullable */
  overallInsight?: string | null;
}

export interface UserProfile {
  id: number;
  nickname: string;
  /** @nullable */
  spendingGoal?: number | null;
  onboardingCompleted: boolean;
  createdAt?: string;
}

export interface UserProfileInput {
  nickname?: string;
  /** @nullable */
  spendingGoal?: number | null;
}

export interface OnboardingInput {
  nickname: string;
  /** @nullable */
  spendingGoal?: number | null;
}

export type ListSpendingRecordsParams = {
limit?: number;
offset?: number;
};