# lib/api-zod/src/generated/types/emotionalAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionalAnalysis {
  /** Identified spending pattern in Korean */
  spendingPattern: string;
  /** Emotional trigger analysis in Korean */
  emotionalTrigger: string;
  /** Personalized advice in Korean */
  personalizedAdvice: string;
}