# lib/api-zod/src/generated/types/onboardingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingInput {
  nickname: string;
  /** @nullable */
  spendingGoal?: number | null;
}