# lib/api-zod/src/generated/types/spendingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SpendingInput {
  /** Amount in Korean Won (KRW) */
  amount: number;
  /** Spending category e.g. 음식, 쇼핑, 오락, 뷰티, 교통, 카페 */
  category: string;
  /** Emotion at time of purchase e.g. 스트레스, 슬픔, 기쁨, 외로움, 무료함, 보상 */
  emotion: string;
  /** Free-text reason for spending */
  reason: string;
  /**
     * Optional additional note
     * @nullable
     */
  note?: string | null;
}