# lib/api-zod/src/generated/types/spendingRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionalAnalysis } from './emotionalAnalysis';

export interface SpendingRecord {
  id: number;
  amount: number;
  category: string;
  emotion: string;
  reason: string;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
  analysis?: EmotionalAnalysis;
}