# lib/api-zod/src/generated/types/spendingSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { BreakdownItem } from './breakdownItem';
import type { WeeklyTrendItem } from './weeklyTrendItem';

export interface SpendingSummary {
  totalAmount: number;
  totalCount: number;
  topEmotion: string;
  topCategory: string;
  emotionBreakdown: BreakdownItem[];
  categoryBreakdown: BreakdownItem[];
  weeklyTrend: WeeklyTrendItem[];
  /** @nullable */
  overallInsight?: string | null;
}