# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FLONE - Emotional Spending Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  nickname: string;
  /** @nullable */
  spendingGoal?: number | null;
  onboardingCompleted: boolean;
  createdAt?: Date;
}